feat: add DateOfBirthBadge and AffiliationBadge components for improved Etrian display
